32bit eeprom size (#230)
* add commitNonVolatileMemory overload to batch write chars n times to flash

* changed uint8_t getEepromBuffer(uint16_t size) to uint8_t getEepromBuffer(uint32_t size) to allow eeprom > 65k

* reworked PR #230

+++ src/cc1310_platform.cpp
@@ -495,7 +495,7 @@ void CC1310Platform::init()
     setupNVS();
 }
 
-uint8_t* CC1310Platform::getEepromBuffer(uint16_t size)
+uint8_t* CC1310Platform::getEepromBuffer(uint32_t size)
 {
     if(size > KNX_FLASH_SIZE)
     {
